- Merge code from file_upload.py into dom.py - maintain the same pattern
