+++ pydoll/commands/dom.py
@@ -1,5 +1,6 @@
 import copy
-from typing import Literal
+from pathlib import Path
+from typing import List, Literal, Optional, Union
 
 from pydoll.commands.runtime import RuntimeCommands
 from pydoll.constants import By, Scripts
@@ -42,6 +43,7 @@ class DomCommands:
         'method': 'DOM.scrollIntoViewIfNeeded',
         'params': {},
     }
+    SET_FILE_INPUT_FILES_TEMPLATE = {'method': 'DOM.setFileInputFiles', 'params': {}}
 
     @classmethod
     def scroll_into_view(cls, object_id: str) -> dict:
@@ -317,3 +319,61 @@ class DomCommands:
                 to make it relative.
         """
         return f'.{xpath}' if not xpath.startswith('.') else xpath
+
+    @staticmethod
+    def _ensure_file_exists(files: Union[str, Path, List[Union[str, Path]]], missing_ok: bool = False) -> List[str]:
+        """
+        Ensures that the file exists.
+
+        Args:
+            files (Union[str, Path, List[Union[str, Path]]]): Files to check.
+            missing_ok (bool): If True, skips the file existence check.
+
+        Returns:
+            List[str]: List of file paths.
+        """
+        if isinstance(files, str):
+            files = [Path(files).absolute()]
+        elif isinstance(files, Path):
+            files = [files]
+
+        _has_ensure_files = []
+        for filepath in files:
+            if isinstance(filepath, str):
+                filepath = Path(filepath).absolute()
+            if missing_ok is False and filepath.is_file() is False:
+                raise FileExistsError(f"{filepath} does not exist.")
+            _has_ensure_files.append(str(filepath))
+
+        return _has_ensure_files
+
+    @classmethod
+    def upload_files(
+            cls,
+            files: Union[str, Path, List[Union[str, Path]]],
+            object_id: Optional[str] = None,
+            backend_node_id: Optional[int] = None,
+            missing_ok: bool = False
+    ) -> dict:
+        """
+        Sets the value of the file input to these file paths or files.
+
+        Args:
+            files (Union[str, Path, List[Union[str, Path]]]): Files to upload.
+            object_id (Optional[str]): JavaScript object id of the node wrapper.
+            backend_node_id (Optional[int]): Identifier of the backend node.
+            missing_ok (bool): If True, skips the file existence check.
+
+        Returns:
+            dict: The CDP command to set the file input files.
+        """
+        command = copy.deepcopy(cls.SET_FILE_INPUT_FILES_TEMPLATE)
+        if object_id is None and backend_node_id is None:
+            raise ValueError("Either object_id or backend_node_id is required.")
+        if object_id is not None:
+            command['params']['objectId'] = object_id
+        if backend_node_id is not None:
+            command['params']['backendNodeId'] = backend_node_id
+        command['params']['files'] = cls._ensure_file_exists(files, missing_ok)
+
+        return command
